requirements.txt
lightgbm>=4.0.0
optuna>=3.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
joblib>=1.3.0
shap>=0.42.0
